# apps/api/openapi.snapshot.json
{
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "description": "Machine-readable error code (e.g. 'invalid_body', 'unauthenticated').",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "JWT (dashboard sessions) or `furan_pat_*` (SDK/CI). Both arrive on `Authorization: Bearer ...`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Visual regression testing API. See https://github.com/qasecret/furan-monorepo for self-host docs.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Furan API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/storage/{key}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Image bytes. Content-Type is sniffed (image/png, image/webp, or image/jpeg)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Key not found, malformed, or storage backend error"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "Authenticated proxy for stored image bytes. Key is sha256 hex of content (content-addressed storage).",
        "tags": [
          "storage"
        ]
      }
    },
...
    "/users": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "createdAt": {
                        "type": "string"
                      },
                      "email": {
                        "format": "email",
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "isActive": {
                        "type": "boolean"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "role": {
                        "enum": [
                          "admin",
                          "editor",
                          "guest"
                        ],
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "email",
                      "firstName",
                      "lastName",
                      "role",
                      "isActive",
                      "createdAt",
                      "updatedAt"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Paginated user list"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid query"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not admin"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "List users (admin only)",
        "tags": [
          "users-admin"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "firstName": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "lastName": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 8,
                    "type": "string"
                  },
                  "role": {
                    "default": "editor",
                    "enum": [
                      "admin",
                      "editor",
                      "guest"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "firstName",
                  "lastName"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "type": "string"
                    },
                    "email": {
                      "format": "email",
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "role": {
                      "enum": [
                        "admin",
                        "editor",
                        "guest"
                      ],
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "firstName",
                    "lastName",
                    "role",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "User created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not admin"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Email already taken"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "Create a user (admin only)",
        "tags": [
          "users-admin"
        ]
      }
    },
...
    "/users/{id}": {
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "firstName": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "lastName": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "role": {
                    "enum": [
                      "admin",
                      "editor",
                      "guest"
                    ],
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "type": "string"
                    },
                    "email": {
                      "format": "email",
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "role": {
                      "enum": [
                        "admin",
                        "editor",
                        "guest"
                      ],
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "firstName",
                    "lastName",
                    "role",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Updated user record"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid body or attempted self-disable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not admin"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": [
            ]
          }
        ],
        "summary": "Update a user (admin only). Cannot disable self.",
        "tags": [
          "users-admin"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}